+++ ui/app/components/network.js
@@ -36,9 +36,6 @@ Network.prototype.render = function () {
   } else if (providerName === 'mainnet') {
     hoverText = 'Main Ethereum Network'
     iconName = 'ethereum-network'
-  } else if (providerName === 'classic') {
-    hoverText = 'Ethereum Classic Network'
-    iconName = 'classic-network'
   } else if (parseInt(networkNumber) === 2) {
     hoverText = 'Morden Test Network'
     iconName = 'morden-test-network'
@@ -65,15 +62,6 @@ Network.prototype.render = function () {
                   color: '#039396',
                 }},
               'Etherum Main Net'),
-            ])
-          case 'classic-network':
-            return h('.network-indicator', [
-              h('.menu-icon.hollow-diamond'),
-              h('.network-name', {
-                style: {
-                  color: '#039396',
-                }},
-              'Etherum Classic'),
             ])
           case 'morden-test-network':
             return h('.network-indicator', [